SZ Implementations of DL algorithm
