+++ README.md
@@ -1,3 +1,3 @@
 # string.distance
-Library to support distances between strings according to several algorithms
+Library to support implementations for the calculation of distances between strings according to several algorithms
