+++ algorithms/README.md
@@ -1,0 +1,6 @@
+Library to support Optimal Alignment
+Library to support Damerau Levenshtein
+Library to support Needleman - Wunsch
+Library to support Smith - Waterman algorithm
+Interfaces for string distances algorithm
+
