+++ algorithms/dameraulevenshtein.go
@@ -1,0 +1,8 @@
+package distances
+
+type DamLev struct {
+}
+
+func (dist DamLev) CalculateDistance(fromString string, toString string) int {
+	return 0
+}
