+++ algorithms/stringdistance.go
@@ -1,0 +1,5 @@
+package distances
+
+type StringDistance interface {
+	CalculateDistance(fromString string, toString string) int
+}
