--- dameraulevenshtein/README.md
@@ -1,1 +1,0 @@
-Library to support Damerau Levenshtein algorithm
