--- dameraulevenshtein/distances.go
@@ -1,11 +1,0 @@
-package dameraulevenshtein
-
-import (
-)
-
-type DamLev struct {
-}
-
-func (dist DamLev) calculateDistance(fromString string, toString string) int {
-	return 0
-}
