--- smithwaterman/README.md
@@ -1,1 +1,0 @@
-Library to support Smith - Waterman algorithm
