--- stringinterface/README.md
@@ -1,1 +1,0 @@
-Interfaces for string distances algorithm
