--- stringinterface/stringdistance.go
@@ -1,7 +1,0 @@
-package stringinterface
-
-import ()
-
-type StringDistance interface {
-	calculateDistance(fromString string, toString string) int
-}
